日志对齐，msg对齐

+++ frameworks/native/appkit/ability_runtime/context/js_context_utils.cpp
@@ -462,7 +462,7 @@ napi_value JsBaseContext::OnGetGroupDir(napi_env env, NapiCallbackInfo& info)
     NapiAsyncTask::ExecuteCallback execute = [context = context_, groupId, &path, innerErrCode]() {
         auto completeContext = context.lock();
         if (!completeContext) {
-            *innerErrCode = static_cast<int32_t>(ERR_ABILITY_RUNTIME_EXTERNAL_CONTEXT_NOT_EXIST);
+            *innerErrCode = ERR_ABILITY_RUNTIME_EXTERNAL_CONTEXT_NOT_EXIST;
             return;
         }
         path = completeContext->GetGroupDir(groupId);
